+++ src/components/select/option/option.component.js
@@ -3,9 +3,17 @@ import PropTypes from "prop-types";
 import StyledOption from "./option.style";
 
 const Option = React.forwardRef(
-  ({ text, children, onSelect, value, isHighlighted, hidden }, ref) => {
+  (
+    { text, children, onSelect, value, isHighlighted, hidden, onClick },
+    ref
+  ) => {
     function handleClick() {
-      onSelect({ text, value });
+      if (!onClick) {
+        onSelect({ text, value });
+      } else {
+        onSelect();
+        onClick({ target: { text, value } });
+      }
     }
 
     return (
@@ -31,6 +39,11 @@ Option.propTypes = {
   children: PropTypes.node,
   /** The option's invisible internal value */
   value: PropTypes.oneOfType([PropTypes.object, PropTypes.string]).isRequired,
+  /**
+   * @private
+   * @ignore
+   * Callback to return value when the element is clicked (prop added by the SelectList component) */
+  onClick: PropTypes.func,
   /**
    * @private
    * @ignore